refactors structure of views and how routing works
