+++ src/App.js
@@ -3,14 +3,15 @@ import {
   BrowserRouter as Router,
   Switch,
   Route,
-  Link
+  Link,
+  Redirect
 } from "react-router-dom";
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub, faTwitter, faInstagram, faLinkedin } from '@fortawesome/free-brands-svg-icons';
-import Home from './views/Home';
-import Portfolio from './views/Portfolio';
+import { Home } from './views/home';
+import { Portfolio } from './views/portfolio';
 import ScrollArrow from './components/scrollArrow';
 import Menu from './components/menu';
 
@@ -26,12 +27,11 @@ export default function Site() {
       </div>
 
       <Switch>
+        <Route exact path="/Home" component={Home} />
         <Route exact path="/">
-          <Home />
+          <Redirect to="/Home" />
         </Route>
-        <Route path="/portfolio">
-          <Portfolio />
-        </Route>
+        <Route exact path="/portfolio" component={Portfolio} />
         <Route path="/articles">
           <Articles />
         </Route>
