+++ src/views/portfolio/TemplateCaseStudy.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+// import { Link } from "react-router-dom";
+
+class CaseStudy extends Component {
+  render() {
+    return (
+      <div className="light feature feature-width full-height">
+        case study template
+      </div>
+    )
+  }
+}
+
+export default CaseStudy;
