+++ src/views/portfolio/index.js
@@ -1,0 +1,1 @@
+export { default as Portfolio } from './Portfolio';
